+++ src/components/Checkbox/test.js
@@ -7,10 +7,14 @@ it('Checkbox renders correctly', () => {
 
   const tree = renderer.create(
     <>
+      <Checkbox indeterminate onChange={changeState} title='Indeterminate'>Checked</Checkbox>
+      <Checkbox checked indeterminate onChange={changeState} >Checked</Checkbox>
       <Checkbox checked onChange={changeState} title='Initial'>Checked</Checkbox>
       <Checkbox onChange={changeState}>Initial</Checkbox>
-      <Checkbox checked onChange={changeState} disabled>Checked disabled</Checkbox>
-      <Checkbox onChange={changeState} disabled>Disabled</Checkbox>
+      <Checkbox disabled checked indeterminate onChange={changeState}>Checked disabled</Checkbox>
+      <Checkbox disabled indeterminate onChange={changeState}>Disabled</Checkbox>
+      <Checkbox disabled checked onChange={changeState}>Checked disabled</Checkbox>
+      <Checkbox disabled onChange={changeState}>Disabled</Checkbox>
     </>
   ).toJSON();