+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Improve Icon's readme.
 
+- Add indeterminate state to Checkbox.
+
 ## [0.17.0] - 2020-05-25
 
 - Testing with jest snapshots